test_data/project1/requirements.txt
attrs==20.1.0
moreorless
